# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: "Planty API"
  description: "API for communication between frontend, microcontroller and backend."
  version: 0.0.4
servers:
  - url: "{protocol}://{host}:{port}/{basePath}"
    description: Production server
    variables:
      protocol:
        enum:
          - "http"
          - "https"
        default: https
        description: Protocol to use for communication.
      host:
        default: "localhost"
        description: Host to use for communication.
      port:
        default: "443"
        description: Port to use for communication.
      basePath:
        default: ""
        description: Base path to use for communication.
paths:
...
  /sensor:
    get:
      summary: "Get all sensors."
      description: "Retrieves all available sensors."
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/sensors"
        "503":
          description: "The service is unavailable."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
    post:
      summary: "Post a sensor."
      description: "Create a new sensor."
      requestBody:
        description: "The sensor to be created."
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/sensorPayload"
      responses:
        "201":
          description: "Created"
        "400":
          description: "The given sensor was invalid."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        "503":
          description: "The service is unavailable."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
  /sensor/{sensorId}:
    parameters:
      - name: sensorId
        description: "ID of the sensor to access."
        in: path
        required: true
        schema:
          type: string
    get:
      summary: "Get a single sensor."
      description: "Retrieves a single sensor based on a given ID."
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/sensor"
        "404":
          description: "No sensor with the given ID was found."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        "500":
          description: "sensorId could not be parsed to an integer value."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        "503":
          description: "The service is unavailable."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
    delete:
      summary: "Delete a single sensor."
      description: "Deletes a single sensor based on a given ID."
      responses:
        "200":
          description: "OK"
        "404":
          description: "No sensor with the given ID was found."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        "500":
          description: "sensorId could not be parsed to an integer value."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        "503":
          description: "The service is unavailable."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
    put:
      summary: "Put a single sensor."
      description: "Updates a sensors configuration."
      requestBody:
        description: "The new sensor configuration."
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/sensorConfigurationPayload"
      responses:
        "200":
          description: "OK"
        "400":
          description: "The given sensor was invalid."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        "404":
          description: "No sensor with the given ID was found."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        "500":
          description: "sensorId could not be parsed to an integer value."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        "503":
          description: "The service is unavailable."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
  /sensor/by-hardware-id/{hardwareId}:
    parameters:
      - name: hardwareId
        description: "Specific hardware ID of the sensor to access."
        in: path
        required: true
        schema:
          type: string
    get:
      summary: "Get a single sensor."
      description: "Retrieves a single sensor based on a given hardware specific ID."
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/sensor"
        "404":
          description: "No sensor with the given hardware specific ID was found."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        "503":
          description: "The service is unavailable."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
components:
  schemas:
...
    errorMessage:
      description: "A json object containing a text message describing an error."
      type: object
      properties:
        message:
          description: "A message describing an error."
          type: string
          maxLength: 256
          example: "The database is not responding."
      example: '{ "message": "The database is not responding." }'
    # Basic schemas
    timestamp:
      description: "A timestamp containing date and time."
      type: string
      format: date-time
...
    entityId:
      description: "The ID schema used for most database entities."
      type: integer
      minimum: 0
      example: 420
    hardwareId:
      description: "The hardware specific ID of a sensor."
      type: string
      maxLength: 50
      example: "ESP32-420"
    entityName:
      description: "A name of an entity"
      type: string
      maxLength: 50
      example: "Sensor 1"
    entityDescription:
      description: "A description of an entity"
      type: string
      maxLength: 50
      example: "The blue ESP32 sensor"
    # Moisture Record-related schemas
...
    sensorSleepTimeout:
      description: "Timeout in minutes for the sensor to sleep in between taking measurements"
      type: integer
    sensorHumidityScalingFrom:
      description: "The minimum value used for scaling the raw sensor values."
      type: integer
    sensorHumidityScalingTo:
      description: "The maximum value used for scaling the raw sensor values."
      type: integer
    sensor:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/entityId"
        hardwareId:
          $ref: "#/components/schemas/hardwareId"
        name:
          $ref: "#/components/schemas/entityName"
        description:
          $ref: "#/components/schemas/entityDescription"
        humidityScalingFrom:
          $ref: "#/components/schemas/sensorHumidityScalingFrom"
        humidityScalingTo:
          $ref: "#/components/schemas/sensorHumidityScalingTo"
        lastCalibration:
          $ref: "#/components/schemas/timestamp"
        sleepTimeout:
          $ref: "#/components/schemas/sensorSleepTimeout"
    sensors:
      type: array
      items:
        $ref: "#/components/schemas/sensor"
    sensorPayload:
      type: object
      properties:
        name:
          $ref: "#/components/schemas/entityName"
        description:
          $ref: "#/components/schemas/entityDescription"
    sensorConfigurationPayload:
      type: object
      properties:
        humidityScalingFrom:
          $ref: "#/components/schemas/sensorHumidityScalingFrom"
        humidityScalingTo:
          $ref: "#/components/schemas/sensorHumidityScalingTo"
        sleepTimeout:
          $ref: "#/components/schemas/sensorSleepTimeout"